Note about using ... for testing error messages
I found dealing with callstacks in error messages pretty confusing when I first started using doctests. I didn't see anything in the README, but later on I saw the suggestion @soenkehahn  made on #136 and it seems to work perfectly.

Hopefully adding this to the README will help other people in the same boat find the solution faster :).